Fix out-of-bound read in console.c
There are multiple types used in parameter list of option command, but
the address of them are all casted to integer pointer and dereferenced
to int in do_option_cmd.

We work around this issue by declaring a void pointer instead of int
pointer for valp in struct PELE, and append valsize indicating the
size of object under valp as a member of struct PELE. Then use
generic_plist_getter() to get the actual value for all report()
functions in console.c

This problem was addressed by Address Sanitizer.